Replace unsigned long with appropriate types (#70)
* Replace unsigned long with appropriate types

This commit aims to replace the generic unsigned long types in the
MPI backend with more descriptive types. Most importantly,
Pyxis related types are now guaranteed to be 64-bit.

* Uniform formatting of for loops

+++ src/virtual_memory/shm.cpp
@@ -52,7 +52,7 @@ namespace argo {
 			struct statvfs b;
 			statvfs("/dev/shm", &b);
 			avail = b.f_bavail * b.f_bsize;
-			if(avail > static_cast<unsigned long>(ARGO_SIZE_LIMIT)) {
+			if(avail > static_cast<std::size_t>(ARGO_SIZE_LIMIT)) {
 				avail = ARGO_SIZE_LIMIT;
 			}
 			std::string filename = "/argocache" + std::to_string(getpid());
